Realization/Backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart

torch
numpy
pyyaml
regex

pandas
pyarrow
fastparquet